@@ -1,14 +1,15 @@
-@testable import _ConnectionPoolModule
+import _ConnectionPoolModule
 import DequeModule
+import NIOConcurrencyHelpers
 
 @available(macOS 13.0, iOS 16.0, tvOS 16.0, watchOS 9.0, *)
-final class MockConnectionFactory<Clock: _Concurrency.Clock>: Sendable where Clock.Duration == Duration {
-    typealias ConnectionIDGenerator = _ConnectionPoolModule.ConnectionIDGenerator
-    typealias Request = ConnectionRequest<MockConnection>
-    typealias KeepAliveBehavior = MockPingPongBehavior
-    typealias MetricsDelegate = NoOpConnectionPoolMetrics<Int>
-    typealias ConnectionID = Int
-    typealias Connection = MockConnection
+public final class MockConnectionFactory<Clock: _Concurrency.Clock>: Sendable where Clock.Duration == Duration {
+    public typealias ConnectionIDGenerator = _ConnectionPoolModule.ConnectionIDGenerator
+    public typealias Request = ConnectionRequest<MockConnection>
+    public typealias KeepAliveBehavior = MockPingPongBehavior
+    public typealias MetricsDelegate = NoOpConnectionPoolMetrics<Int>
+    public typealias ConnectionID = Int
+    public typealias Connection = MockConnection
 
     let stateBox = NIOLockedValueBox(State())
 
@@ -20,18 +21,33 @@ final class MockConnectionFactory<Clock: _Concurrency.Clock>: Sendable where Clo
         var runningConnections = [ConnectionID: Connection]()
     }
 
-    var pendingConnectionAttemptsCount: Int {
+    let autoMaxStreams: UInt16?
+
+    public init(autoMaxStreams: UInt16? = nil) {
+        self.autoMaxStreams = autoMaxStreams
+    }
+
+    public var pendingConnectionAttemptsCount: Int {
         self.stateBox.withLockedValue { $0.attempts.count }
     }
 
-    var runningConnections: [Connection] {
+    public var runningConnections: [Connection] {
         self.stateBox.withLockedValue { Array($0.runningConnections.values) }
     }
 
-    func makeConnection(
+    public func makeConnection(
         id: Int,
         for pool: ConnectionPool<MockConnection, Int, ConnectionIDGenerator, some ConnectionRequestProtocol, Int, MockPingPongBehavior<MockConnection>, NoOpConnectionPoolMetrics<Int>, Clock>
     ) async throws -> ConnectionAndMetadata<MockConnection> {
+        if let autoMaxStreams = self.autoMaxStreams {
+            let connection = MockConnection(id: id)
+            Task {
+                try? await connection.signalToClose
+                connection.closeIfClosing()
+            }
+            return .init(connection: connection, maximalStreamsOnConnection: autoMaxStreams)
+        }
+
         // we currently don't support cancellation when creating a connection
         let result = try await withCheckedThrowingContinuation { (checkedContinuation: CheckedContinuation<(MockConnection, UInt16), any Error>) in
             let waiter = self.stateBox.withLockedValue { state -> (CheckedContinuation<(ConnectionID, CheckedContinuation<(MockConnection, UInt16), any Error>), Never>)? in
@@ -52,7 +68,7 @@ final class MockConnectionFactory<Clock: _Concurrency.Clock>: Sendable where Clo
     }
 
     @discardableResult
-    func nextConnectAttempt(_ closure: (ConnectionID) async throws -> UInt16) async rethrows -> Connection {
+    public func nextConnectAttempt(_ closure: (ConnectionID) async throws -> UInt16) async rethrows -> Connection {
         let (connectionID, continuation) = await withCheckedContinuation { (continuation: CheckedContinuation<(ConnectionID, CheckedContinuation<(MockConnection, UInt16), any Error>), Never>) in
             let attempt = self.stateBox.withLockedValue { state -> (ConnectionID, CheckedContinuation<(MockConnection, UInt16), any Error>)? in
                 if let attempt = state.attempts.popFirst() {
