@@ -29,6 +29,12 @@ struct TinyFastSequence<Element>: Sequence {
             self.base = .none(reserveCapacity: 0)
         case 1:
             self.base = .one(collection.first!, reserveCapacity: 0)
+        case 2:
+            self.base = .two(
+                collection.first!,
+                collection[collection.index(after: collection.startIndex)],
+                reserveCapacity: 0
+            )
         default:
             if let collection = collection as? Array<Element> {
                 self.base = .n(collection)
@@ -46,7 +52,7 @@ struct TinyFastSequence<Element>: Sequence {
         case 1:
             self.base = .one(max2Sequence.first!, reserveCapacity: 0)
         case 2:
-            self.base = .n(Array(max2Sequence))
+            self.base = .two(max2Sequence.first!, max2Sequence.second!, reserveCapacity: 0)
         default:
             fatalError()
         }
@@ -169,7 +175,7 @@ struct TinyFastSequence<Element>: Sequence {
 
             case .n(let array):
                 if self.index < array.endIndex {
-                    defer { self.index += 1}
+                    defer { self.index += 1 }
                     return array[self.index]
                 }
                 return nil
