@@ -36,13 +36,20 @@ extension String: PostgresDecodable {
             // we can force unwrap here, since this method only fails if there are not enough
             // bytes available.
             self = buffer.readString(length: buffer.readableBytes)!
+
         case (_, .uuid):
             guard let uuid = try? UUID(from: &buffer, type: .uuid, format: format, context: context) else {
                 throw PostgresDecodingError.Code.failure
             }
             self = uuid.uuidString
+
         default:
-            throw PostgresDecodingError.Code.typeMismatch
+            // We should eagerly try to convert any datatype into a String. For example the oid
+            // for ltree isn't static. For this reason we should just try to convert anything.
+            guard let string = buffer.readString(length: buffer.readableBytes) else {
+                throw PostgresDecodingError.Code.typeMismatch
+            }
+            self = string
         }
     }
 }
