@@ -1,4 +1,5 @@
 import Crypto
+import _CryptoExtras
 import Foundation
 
 extension UInt8 {
@@ -292,7 +293,7 @@ internal struct SHA256: SASLAuthenticationMechanism {
     ///               authenticating user. If the closure throws, authentication
     ///               immediately fails with the thrown error.
     internal init(username: String, password: @escaping () throws -> String) {
-        self._impl = .init(username: username, passwordGrabber: { _ in try (Array(password().data(using: .utf8)!), []) }, bindingInfo: .unsupported)
+        self._impl = .init(username: username, passwordGrabber: { _ in try (Array(password().utf8), []) }, bindingInfo: .unsupported)
     }
     
     /// Set up a server-side `SCRAM-SHA-256` authentication.
@@ -338,7 +339,7 @@ internal struct SHA256_PLUS: SASLAuthenticationMechanism {
     ///   - channelBindingData: The appropriate data associated with the RFC5056
     ///                         channel binding specified.
     internal init(username: String, password: @escaping () throws -> String, channelBindingName: String, channelBindingData: [UInt8]) {
-        self._impl = .init(username: username, passwordGrabber: { _ in try (Array(password().data(using: .utf8)!), []) }, bindingInfo: .bind(channelBindingName, channelBindingData))
+        self._impl = .init(username: username, passwordGrabber: { _ in try (Array(password().utf8), []) }, bindingInfo: .bind(channelBindingName, channelBindingData))
     }
     
     /// Set up a server-side `SCRAM-SHA-256` authentication.
@@ -466,8 +467,8 @@ fileprivate final class SASLMechanism_SCRAM_SHA256_Common {
         // TODO: Perform `Normalize(password)`, aka the SASLprep profile (RFC4013) of stringprep (RFC3454)
         
         // Calculate `AuthMessage`, `ClientSignature`, and `ClientProof`
-        let saltedPassword = Hi(string: password, salt: serverSalt, iterations: serverIterations)
-        let clientKey = HMAC<SHA256>.authenticationCode(for: "Client Key".data(using: .utf8)!, using: .init(data: saltedPassword))
+        let saltedPassword = try Hi(string: password, salt: serverSalt, iterations: serverIterations)
+        let clientKey = HMAC<SHA256>.authenticationCode(for: Data("Client Key".utf8), using: saltedPassword)
         let storedKey = SHA256.hash(data: Data(clientKey))
         var authMessage = firstMessageBare; authMessage.append(.comma); authMessage.append(contentsOf: message); authMessage.append(.comma); authMessage.append(contentsOf: clientFinalNoProof)
         let clientSignature = HMAC<SHA256>.authenticationCode(for: authMessage, using: .init(data: storedKey))
@@ -485,9 +486,11 @@ fileprivate final class SASLMechanism_SCRAM_SHA256_Common {
         var clientFinalMessage = clientFinalNoProof; clientFinalMessage.append(.comma)
         guard let proofPart = SCRAMMessageParser.serialize([.p(Array(clientProof))]) else { throw SASLAuthenticationError.genericAuthenticationFailure }
         clientFinalMessage.append(contentsOf: proofPart)
-        
+
+        let saltedPasswordBytes = saltedPassword.withUnsafeBytes { [UInt8]($0) }
+
         // Save state and send
-        self.state = .clientSentFinalMessage(saltedPassword: saltedPassword, authMessage: authMessage)
+        self.state = .clientSentFinalMessage(saltedPassword: saltedPasswordBytes, authMessage: authMessage)
         return .continue(response: clientFinalMessage)
     }
     
@@ -501,7 +504,7 @@ fileprivate final class SASLMechanism_SCRAM_SHA256_Common {
         switch incomingAttributes.first {
             case .v(let verifier):
                 // Verify server signature
-                let serverKey = HMAC<SHA256>.authenticationCode(for: "Server Key".data(using: .utf8)!, using: .init(data: saltedPassword))
+                let serverKey = HMAC<SHA256>.authenticationCode(for: Data("Server Key".utf8), using: .init(data: saltedPassword))
                 let serverSignature = HMAC<SHA256>.authenticationCode(for: authMessage, using: .init(data: serverKey))
                 
                 guard Array(serverSignature) == verifier else {
@@ -585,7 +588,7 @@ fileprivate final class SASLMechanism_SCRAM_SHA256_Common {
         guard nonce == repeatNonce else { throw SASLAuthenticationError.genericAuthenticationFailure }
         
         // Compute client signature
-        let clientKey = HMAC<SHA256>.authenticationCode(for: "Client Key".data(using: .utf8)!, using: .init(data: saltedPassword))
+        let clientKey = HMAC<SHA256>.authenticationCode(for: Data("Client Key".utf8), using: .init(data: saltedPassword))
         let storedKey = SHA256.hash(data: Data(clientKey))
         var authMessage = clientBareFirstMessage; authMessage.append(.comma); authMessage.append(contentsOf: serverFirstMessage); authMessage.append(.comma); authMessage.append(contentsOf: message.dropLast(proof.count + 3))
         let clientSignature = HMAC<SHA256>.authenticationCode(for: authMessage, using: .init(data: storedKey))
@@ -604,7 +607,7 @@ fileprivate final class SASLMechanism_SCRAM_SHA256_Common {
         guard storedKey == restoredKey else { throw SCRAMServerError.invalidProof }
         
         // Compute server signature
-        let serverKey = HMAC<SHA256>.authenticationCode(for: "Server Key".data(using: .utf8)!, using: .init(data: saltedPassword))
+        let serverKey = HMAC<SHA256>.authenticationCode(for: Data("Server Key".utf8), using: .init(data: saltedPassword))
         let serverSignature = HMAC<SHA256>.authenticationCode(for: authMessage, using: .init(data: serverKey))
         
         // Generate a `server-final-message`
@@ -640,19 +643,12 @@ fileprivate final class SASLMechanism_SCRAM_SHA256_Common {
   HMAC() == output length of H().
   ````
 */
-private func Hi(string: [UInt8], salt: [UInt8], iterations: UInt32) -> [UInt8] {
-    let key = SymmetricKey(data: string)
-    var Ui = HMAC<SHA256>.authenticationCode(for: salt + [0x00, 0x00, 0x00, 0x01], using: key) // salt + 0x00000001 as big-endian
-    var Hi = Array(Ui)
-    
-    Hi.withUnsafeMutableBytes { Hibuf -> Void in
-        for _ in 2...iterations {
-            Ui = HMAC<SHA256>.authenticationCode(for: Data(Ui), using: key)
-            
-            Ui.withUnsafeBytes { Uibuf -> Void in
-                for i in 0..<Uibuf.count { Hibuf[i] ^= Uibuf[i] }
-            }
-        }
-    }
-    return Hi
+private func Hi(string: [UInt8], salt: [UInt8], iterations: UInt32) throws -> SymmetricKey {
+    try KDF.Insecure.PBKDF2.deriveKey(
+        from: string,
+        salt: salt,
+        using: .sha256,
+        outputByteCount: 32,
+        unsafeUncheckedRounds: Int(iterations)
+    )
 }
