@@ -1,7 +1,8 @@
 @testable import _ConnectionPoolModule
+import _ConnectionPoolTestUtils
 import Atomics
-import XCTest
 import NIOEmbedded
+import XCTest
 
 @available(macOS 13.0, iOS 16.0, tvOS 16.0, watchOS 9.0, *)
 final class ConnectionPoolTests: XCTestCase {
